Edit sizeGrid to prevent style tag duplication
If the user changed the grid size multiple times in a session, there
would be multiple style tags in the head, which functionally worked as
the last appended style has highest inheritance but might bloat the head
to an unreasonable amount. Now, previous style tag will be removed
before the new one is applied.